refactor: split semantic check
